+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix potential bugs related to system settings
+
 ## [1.4.0] - 2022-10-19
 
 ### Added